fixes #216 Removed call to MeshNormalSpec::CheckNormals(). Not necessary since we explicity set all normals.
Also, we now call the faster SetAllExplicit method rather than a per vertex call.

Finally, the MNMesh::BuildsNormals() instead MNMesh::CheckNormals(). We just set all normals, no need to check if they need to be rebuilt.
We only want build the RVert array (although maybe we can remove this call in the future if normal queries are not needed)